MCC-69506 - Selenium Grid and SpecRun support for Parallel/Cross browser testing(NRTM)

+++ Medidata.RBT.PageObjects.Rave/SharedRaveObjects/GlobalConfiguration.cs
@@ -13,6 +13,7 @@ using Medidata.RBT.PageObjects.Rave.Configuration;
 using Medidata.RBT.PageObjects.Rave.PDF;
 using Medidata.RBT.SharedObjects;
 using System.IO;
+using Medidata.RBT.ConfigurationHandlers;
 
 namespace Medidata.RBT.PageObjects.Rave.SharedRaveObjects
 {
@@ -79,8 +80,10 @@ namespace Medidata.RBT.PageObjects.Rave.SharedRaveObjects
         public void DeleteSelf()
         {
             File.Delete(UniqueFileLocation);
-            GlobalConfiguration clearedConfig = SeedingContext.GetExistingFeatureObjectOrMakeNew<GlobalConfiguration>(DEFAULT_CONFIGURATION, () => new GlobalConfiguration(DEFAULT_CONFIGURATION));
-            File.Delete(clearedConfig.UniqueFileLocation);
+            GlobalConfiguration clearedConfig = SeedingContext.TryGetExistingFeatureObject<GlobalConfiguration>(DEFAULT_CONFIGURATION);
+
+            if (clearedConfig != null)
+                File.Delete(clearedConfig.UniqueFileLocation);
         }
     }
 }
